requirements.txt
streamlit
pandas
numpy
scikit-learn
litellm
google-adk
mistral
mistralai
matplotlib
langchain
langchain-mistralai
langchain-core